% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_abundances.R
\name{HotLoadings.combine_abundances}
\alias{HotLoadings.combine_abundances}
\title{Combine abundances between groups}
\usage{
HotLoadings.combine_abundances(top_features, mean_relative_abundances,
  Y_name)
}
\arguments{
\item{top_features}{A data frame containing \code{n_top} features generated by \code{\link{HotLoadings.top_features}}.}

\item{mean_relative_abundances}{A data frame obtained with \code{\link{HotLoadings.mean_relative_abundances}} containings mean relative abundances for each feature.}

\item{Y_name}{A character string indicating which is the variable associated with component specified. Must be a dicothomous variable.}
}
\value{
A data frame containings \code{n_top} feature associated with a specified \code{component}, information about their mean relative abundance in the 2 \code{Y_name} categories, loading values, and proportions useful for plot.
}
\description{
Combine abundances between groups
}
\seealso{
\code{\link{HotLoadings.mean_relative_abundances}} to create mean relative abundances data frame and \code{\link{HotLoadings.plot_loadings}} for plot.
}
