% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich_loadings.R
\name{HotLoadings.heat_map_long}
\alias{HotLoadings.heat_map_long}
\title{Plot Heatmap the long way}
\usage{
HotLoadings.heat_map_long(
  data.splsda,
  top_feature,
  PSOBJ,
  sample_name,
  Y_name,
  feature_names
)
}
\arguments{
\item{data.splsda}{A sPLS-DA object from mixomics.}

\item{top_feature}{A data frame containing \code{n_top} features generated by \code{\link{HotLoadings.top_features}}.}

\item{PSOBJ}{A phyloseq object.}

\item{sample_name}{A character string indicating which is the variable with sample names.}

\item{Y_name}{A character string indicating which is the variable associated with component specified. Must be a dicothomous variable.}

\item{feature_names}{A character vector containing feature names.}
}
\value{
The function plots heatmap for \code{n_top} features in rows and samples in columns.
}
\description{
The user need to previously launch \code{\link{HotLoadings.names}}, \code{\link{HotLoadings.top_features}}, in order to make this function works. Otherwise use \code{\link{HotLoadings.heat_map}}
}
\seealso{
\code{\link{HotLoadings.plot_loadings_long}} or \code{\link{HotLoadings.plot_loadings}} to plot loadings.
}
