% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_relative_abundances.R
\name{HotLoadings.mean_relative_abundances}
\alias{HotLoadings.mean_relative_abundances}
\title{Compute Mean Relative Abundances}
\usage{
HotLoadings.mean_relative_abundances(top_features, PSOBJ, Y_name)
}
\arguments{
\item{top_features}{A data frame containing \code{n_top} features generated by \code{\link{HotLoadings.top_features}}.}

\item{PSOBJ}{A phyloseq object.}

\item{Y_name}{A character string indicating which is the variable associated with component specified. Must be a dicothomous variable.}
}
\value{
A data frame containings \code{n_top} feature associated with a specified \code{component} and information about their mean relative abundance in the 2 \code{Y_name} categories.
}
\description{
Compute Mean Relative Abundances
}
\seealso{
\code{\link{HotLoadings.top_features}} to create top features data frame.
}
