% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{HotLoadings.names}
\alias{HotLoadings.names}
\title{Extract feature names}
\usage{
HotLoadings.names(PSOBJ, format = c("last", "long", "short"))
}
\arguments{
\item{PSOBJ}{A phyloseq object.}

\item{format}{"last" if only the deepest taxonomic characterization is needed. "long" if all taxonomic levels are needed. "short" if last 2 known taxonomic levels are needed.}
}
\value{
A data frame containings feature names in wanted format.
}
\description{
Extract feature names
}
\examples{
library(phyloseq)
data("GlobalPatterns")
feature_names <- HotLoadings.names(PSOBJ = GlobalPatterns, format = "short")
head(feature_names)
}
