% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich_loadings.R
\name{HotLoadings.plot_loadings}
\alias{HotLoadings.plot_loadings}
\title{Plot Loadings the short way}
\usage{
HotLoadings.plot_loadings(PSOBJ, format = c("short", "last", "long"),
  data.splsda, Y_name, component, n_top = 15, offset = 0.005,
  xlim = c(-0.3, 0.3), colors = c("orange", "red", "blue",
  "dark green"))
}
\arguments{
\item{PSOBJ}{A phyloseq object.}

\item{format}{"last" if only the deepest taxonomic characterization is needed. "long" if all taxonomic levels are needed. "short" if last 2 known taxonomic levels are needed.}

\item{data.splsda}{A sPLS-DA object from mixomics.}

\item{Y_name}{A character string indicating which is the variable associated with component specified. Must be a dicothomous variable.}

\item{component}{A character string containing sPLS-DA's component name.}

\item{n_top}{An integer number indicating how many top associated features to keep. (Default = 15)}

\item{offset}{A value indicating the amount of justification from axis for possible plots. (Default = 0.005)}

\item{xlim}{Vector of size 2 containings plot limits for x axis of the graphic (default is -0.3, 0.3).}

\item{colors}{Vector of size 4 containings colors for each class of the graphic.
As negative and positive values of the loading has a precise meaning, color order is important.
If "Condition 1" is positively associated to specified component (positive loading) and "Condition 2" viceversa:
Color1 represents the proportion of negative loading features in "Condition 1" samples;
Color2 represents the proportion of negative loading features in "Condition 2" samples;
Color3 represents the proportion of positive loading features in "Condition 1" samples;
Color4 represents the proportion of positive loading features in "Condition 2" samples;}
}
\value{
The function plots loadings with several informations:
  * \code{n_top} associated feature with component;
  * Relative abundances for each feature in every condition (white numbers);
  * The proportion between relative abundances for each feature in every condition (length of different colors inside each bar);
}
\description{
The user need to be at the end of mixOmics pipeline.
}
\seealso{
\code{\link{HotLoadings.combine_abundances}} to create combined condition abundances data frame and \code{\link{HotLoadings.plot_loadings}} for plot in a single step.
}
