% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich_loadings.R
\name{HotLoadings.plot_loadings_long}
\alias{HotLoadings.plot_loadings_long}
\title{Plot Loadings the long way}
\usage{
HotLoadings.plot_loadings_long(
  top_feature,
  top_feature_combined,
  xlim = c(-0.3, 0.3),
  colors = c("orange", "red", "blue", "dark green")
)
}
\arguments{
\item{top_feature}{A data frame containing \code{n_top} features generated by \code{\link{HotLoadings.top_features}}.}

\item{top_feature_combined}{A data frame containing \code{n_top} features informations and proportion between \code{Y_name} categories. Generated by \code{\link{HotLoadings.combine_abundances}}.}

\item{xlim}{Vector of size 2 containings plot limits for x axis of the graphic (default is -0.3, 0.3).}

\item{colors}{Vector of size 4 containings colors for each class of the graphic.
As negative and positive values of the loading has a precise meaning, color order is important.
If "Condition 1" is positively associated to specified component (positive loading) and "Condition 2" viceversa:
Color1 represents the proportion of negative loading features in "Condition 1" samples;
Color2 represents the proportion of negative loading features in "Condition 2" samples;
Color3 represents the proportion of positive loading features in "Condition 1" samples;
Color4 represents the proportion of positive loading features in "Condition 2" samples;}
}
\value{
The function plots loadings with several informations:
  * \code{n_top} associated feature with component;
  * Relative abundances for each feature in every condition (white numbers);
  * The proportion between relative abundances for each feature in every condition (length of different colors inside each bar);
}
\description{
The user need to previously launch \code{HotLoadings.names}, \code{HotLoadings.top_features}, \code{HotLoadings.mean_relative_abundances} and \code{HotLoadings.combine_abundances} in order to make this function works. Otherwise use \code{\link{HotLoadings.plot_loadings}}
}
\seealso{
\code{\link{HotLoadings.combine_abundances}} to create combined condition abundances data frame and \code{\link{HotLoadings.plot_loadings}} for plot in a single step.
}
