% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_loadings.R
\name{HotLoadings.plot_loadings_simple}
\alias{HotLoadings.plot_loadings_simple}
\title{HotLoadings.plot_loadings_simple
Plot Loadings the simple way}
\usage{
HotLoadings.plot_loadings_simple(
  PSOBJ,
  format = c("short", "last", "long"),
  data.splsda,
  method = "mean",
  contrib = "max",
  Y_name,
  comp,
  ndisplay = 15,
  offset = 0.005,
  colors = NULL
)
}
\arguments{
\item{PSOBJ}{A phyloseq object.}

\item{format}{"last" if only the deepest taxonomic characterization is needed. "long" if all taxonomic levels are needed. "short" if last 2 known taxonomic levels are needed.}

\item{data.splsda}{A sPLS-DA object from mixomics.}

\item{method}{a character set to 'mean' or 'median' indicating the criterion
to assess the contribution. We recommend using median in the case of count
or skewed data.}

\item{contrib}{a character set to 'max' or 'min' indicating if the color of
the bar should correspond to the group with the maximal or minimal
expression levels / abundance.}

\item{Y_name}{A character string indicating which is the variable associated with component specified. Must be a dicothomous variable.}

\item{comp}{integer value indicating the component of interest from the
object.}

\item{ndisplay}{integer indicating how many of the most important variables
are to be plotted (ranked by decreasing weights in each PLS-component).
Useful to lighten a graph.}

\item{offset}{A value indicating the amount of justification from axis for possible plots. (Default = 0.005)}

\item{colors}{Vector of Hexadecimal codes for colors. Must be the same length
of Y levels.}
}
\value{
The function return a ggplot2 loadings object.
}
\description{
HotLoadings.plot_loadings_simple
Plot Loadings the simple way
}
