% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich_loadings.R
\name{HotLoadings}
\alias{HotLoadings}
\title{Plot Loadings and Heatmap in a sexy way}
\usage{
HotLoadings(g_loadings, g_heat_map, rel_widths = c(1, 1))
}
\arguments{
\item{g_loadings}{a ggplot object obtained with \code{\link{HotLoadings.plot_loadings}} or \code{\link{HotLoadings.plot_loadings_long}}}

\item{g_heat_map}{a ggplot object obtained with \code{\link{HotLoadings.heat_map}} or \code{\link{HotLoadings.heat_map_long}}}

\item{rel_widths}{Numerical vector of relative columns widths. For example, in a two-column grid, rel_widths = c(2, 1) would make the first column twice as wide as the second column.}
}
\value{
The function plots heatmap for \code{n_top} features in rows and samples in columns.
}
\description{
The user need to be at the end of mixOmics pipeline.
}
\seealso{
\code{\link{HotLoadings.plot_loadings_long}} or \code{\link{HotLoadings.plot_loadings}} to plot loadings.
}
