% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_features.R
\name{HotLoadings.top_features}
\alias{HotLoadings.top_features}
\title{Select top features}
\usage{
HotLoadings.top_features(data.splsda, feature_names,
  component = c("comp 1"), Y_name, n_top = 15, offset = 0.005,
  order = TRUE)
}
\arguments{
\item{data.splsda}{A sPLS-DA object from mixomics.}

\item{feature_names}{A character vector containing feature names.}

\item{component}{A character string containing sPLS-DA's component name.}

\item{Y_name}{A character string indicating which is the variable associated with component specified. Must be a dicothomous variable.}

\item{n_top}{An integer number indicating how many top associated features to keep. (Default = 15)}

\item{offset}{A value indicating the amount of justification from axis for possible plots. (Default = 0.005)}

\item{order}{A boolean value indicating if the loadings should be separately ordered by positives and negatives (Default = TRUE). If FALSE loadings are ordered using their absolute value.}
}
\value{
A data frame containings \code{n_top} feature associated with \code{component} with information about interesting \code{Y_name} variable.
}
\description{
Select top features
}
\examples{
library(phyloseq)
library(mixOmics)
# diverse.16S = mixOmics::diverse.16S

# the 16S normalised data
# diverse.TSS = diverse.16S$data.TSS
# the outcome
# Y = diverse.16S$bodysite
# unique ID of each individual for multilevel analysis
# sample = diverse.16S$sample

# after several steps and tuning
# select.keepX = c(20,20,20)
# diverse.splsda = splsda(diverse.TSS, Y, ncomp = 3, logratio = 'CLR', multilevel = sample, keepX = select.keepX)

# for feature names we should use a phyloseq object
# diverse.ps <- phyloseq(otu_table(diverse.16S$data.raw,taxa_are_rows = FALSE),sample_data(diverse.16S$indiv),tax_table(diverse.16S$taxonomy))

# feature_names <- HotLoadings.names(PSOBJ = diverse.ps, format = "short")

# HotLoadings.top_features(data.splsda = diverse.splsda,feature_names = feature_names,component = "comp 1",Y_name = "bodysite")
# feature_names <- HotLoadings.names(PSOBJ = diverse.ps, format = "short")
# head(feature_names)
}
\seealso{
\code{\link{HotLoadings.names}} for extraction of \code{feature_names}
}
