% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich_loadings.R
\name{plot_heat_map}
\alias{plot_heat_map}
\title{Plot Heatmap the short way}
\usage{
plot_heat_map(PSOBJ, format = c("short", "last", "long"), data.splsda,
  Y_name, component, n_top = 15, sample_name, offset = 0.005)
}
\arguments{
\item{PSOBJ}{A phyloseq object.}

\item{format}{"last" if only the deepest taxonomic characterization is needed. "long" if all taxonomic levels are needed. "short" if last 2 known taxonomic levels are needed.}

\item{data.splsda}{A sPLS-DA object from mixomics.}

\item{Y_name}{A character string indicating which is the variable associated with component specified. Must be a dicothomous variable.}

\item{component}{A character string containing sPLS-DA's component name.}

\item{n_top}{An integer number indicating how many top associated features to keep. (Default = 15)}

\item{sample_name}{A character string indicating which is the variable with sample names.}

\item{offset}{A value indicating the amount of justification from axis for possible plots. (Default = 0.005)}
}
\value{
The function plots heatmap for \code{n_top} features in rows and samples in columns.
}
\description{
The user need to be at the end of mixOmics pipeline.
}
\seealso{
\code{\link{HotLoadings.plot_loadings_long}} or \code{\link{HotLoadings.plot_loadings}} to plot loadings.
}
